+++ CHANGELOG.md
@@ -36,6 +36,7 @@ Ref: https://keepachangelog.com/en/1.0.0/
 
 ## Unreleased
 
+-(fix) DAPP-59 | packages/i18n 1.0.6 apps/mission 1.0.36 | footer airfoil
 -(fix) DAPP-22 | packages/ui 1.0.1 | storybooks: chips
 
 ## 2.0.11 - 2024-03-13